all: use the term 'mapping' instead of 'transformations'
Use the term 'mapping' instead of 'transformations' when referring to
mappings from properties to other properties.

Note that the term 'transformation' is still used when referring to
Python transformations; also, in the UI the term 'transformations' still
appears in JS code and error messages, and should be updated in future
commits.