apis: make some fixes to the identity solver
Currently the code of the identity solver seems to work, but this is due
to some bugs in the code. This commit fixes that bugs and, as a
consequent, breaks some behaviors that seem to work now but,
technically, was unexpected behaviors.

In future commits, the identity solver will be reviewed and reasonably
re-implemented using the memory state.