apis: improve error handling when serving HTTP requests
This commit standardizes the error handling when serving HTTP requests,
fixing some errors not properly handled, reducing the complexity of the
code and simplify the adding of other handlers in the future.

Note that the behavior has changed: now the `Content-Type:
application/json` header is not always returned (it must be explicitly
set before returning a JSON body), so there are methods which may return
the status code 200 and an empty body.