.github/workflows: remove the action `check_on_commit.yml`
The action described by `check_on_commit.yml` fails because it needs to
access `github.com/open2b/nuts`, which is a private repo.

Considering that we don not vendor dependencies in this repository, this
would work using some other solutions (eg. by adding an SSH key to the
action, etc...), but that is not trivial and may introduce some other
problems.

So, for now, the Action won't run.