apis/types: remove nullability of types
Types nullability was originally implemented in commit
ba5a27de6138b5e79e12867ebaf302c4d4ad0c9c.

This commit removes nullability of types because it will be implemented
later on properties, not on types.

Note that some code that handles the nullability has been kept (in some
cases a TODO has been added) because it will be used again when the
properties will be nullable.