apis: invert keys and values in action properties mapping
Currently, in the map[string]string which represents the action of a
mapping, the map keys represents the "left" properties (the properties
of the incoming event) and the map values represents the "right"
properties (the properties of the mapped event).

This commit inverts the meaning of the keys and values of the mappings
associated to the actions.