admin/src: standardize the buttons usage
This commit standardizes the usage of button icons throughout the app by
removing them when they do not improve the readability (as in a "Save"
button) and by leaving them only when the button is a navigation button
or expresses an higher level concept (as in "Map the properties").

In addition, ellipsis have been added to all the buttons (with the
exception of navigation buttons) that require additional user actions
before executing the operation expressed in their text.