connector: change `OAuth.ExpiresIn` type from `time.Duration` to `int32`
"expires_in" in OAuth is expressed in seconds, so it's better to also
express OAuth.ExpiresIn in seconds instead of nanoseconds. Using
time.Duration leads to errors, and in fact, both the documentation of
OAuth.ExpiresIn is wrong, and the value of ExpiresIn in the Mailchimp
connector was mistakenly expressed in seconds instead of nanoseconds.
This commit also fixes these errors.