apis: associate destination users to actions instead of connections
Currently, we store the destination users to the connections instead of
the actions; this is not correct, since we may have more than one export
action per destination connection, and also makes impossible to remove
the users when an action is removed.

This commit associates the destinations users to the actions instead of
the connections.

Also take a look at the issue: https://github.com/open2b/chichi/issues/179