all: remove `pkg/pgx` and use `https://github.com/jackc/pgx/` v5.3.1
The module `https://github.com/jackc/pgx/` was copy-pasted in our
repository to add a workaround for a bug (not calling the Scan method)
in that module.

Such bug has been solved in `https://github.com/jackc/pgx`, so this
commit uses again the upstream version and remove the copied files from
this repository.