apis/mappings: check enum and regexp while converting
While converting to text, if the destination type has an enum, it checks
whether the value is present in the enum, and if the destination has a
regex, it checks whether the value matches the regular expression.

If the value is not present in the enum or does not match the regex, but
is an empty string and the property is nullable, it converts the value
to nil.