apis: always validate the integer returned by `strconv.Atoi`
`strconv.Atoi(s)` returns `math.MaxInt` if `s` represents an integer,
but it overflows `math.MaxInt`. This commit ensures that this case is
correctly handled.

Note that in case of error, `Atoi` returns either `0` or `math.MaxInt`.
Therefore, there is no need to handle the error returned by `Atoi` if
the resulting number is validated and `0` and `math.MaxInt` are not
considered valid.