test/chichitester: make the tests fail if an action ends with an error
Currently, if an action ends with an error the test proceeds without
altering its flow; this may lead to unexpected test behaviors and
situations which are hard to debug / understand.

This commit changes the support method `WaitActionsToFinish` to make the
test fail in case of an action execution has ended with an error.