test/chichitester: increase the timeout after an HTTP call to Chichi
On some computers, a timeout of 10ms may not be enough to ensure that
the state of Chichi has been updated by the last call to the APIs.

So, this commit increases the timeout from 10ms to 30ms to decrease the
possibilities that such behavior occurs.

See the issue https://github.com/open2b/chichi/issues/193 for more
details about this.