apis/events: change the type of `collectedEvent.Properties`
This commit changes the type of `collectedEvent.Properties` from
`json.RawMessage` to `map[string]any`. This change reduces the cases
where properties need to be converted from `json.RawMessage` to
`map[string]any` and vice versa.

It also fixes the conversion of properties, as properties that were
Number in JSON were serialized as `float64` instead of `json.Number`.