server: stop using the Go default mutex for handling HTTP requests
The Go default mutex, used for handling HTTP requests, due to a bug in
its implementation doesn't handle escaped sequences in the URL path like
"%2F%2F", making impossible to call methods like:

	GET     /api/connections/{id}/complete-path/{path}

while passing a {path} which contains, for example, trailing slashes.

This commit replaces the default mutex with a single HTTP handler that
executes a switch statement, as a workaround for the problem.