apis: fix validation of mappings in `validateActionToSet`
Currently, when mapping is not mandatory (for example for actions that
send events), if the mapping is provided then `validateActionToSet`
returns error.

This behavior is correct when the event type has no schema associated,
but it is wrong when such schema is valid.

So, since the method `validateActionToSet` knows nothing about the event
type (except its name), then the error should be never returned here.