admin: split the effect in `AppProvider`
Currently, in `AppProvider`, when the connections are set as stale they
are refetched from inside an effect.

However, since the connectors are fetched in the same effect, whenever
the connections are stale, the connectors are refetched with them.

This commit fix this issue by splitting the effect in two independent
effects: one dedicated to fetching the connectors and one dedicated to
fetching the connections.