apis: improve action validation
The code for validating mappings has been moved to the beginning of the
`validateActionToSet` method since it performs static validation.
Additionally, the names of the properties used as input and output in
the transformation are now being validated.

In addition, this change makes the filter requires the input schema and
changes the order of the `Action` fields.