admin, apis: add server-side validation for expressions
This commit moves the validation logic of mapping expressions from the client to the server, implementing the HTTP endpoints `validate-expression` and `expressions-properties`.

With the `validate-expression` endpoint, the UI can now validate the expressions entered by the user in the mapping input fields through an HTTP call, without the need to implement additional validation logic in the application. Any error message returned will be displayed below the input field, and attempting to save the action will display the error message "You must fix the errors on the mapping before saving".

The same behavior (with a different error message) has also been implemented for action identifiers and anonymous identifiers.

Furthermore, thanks to the `expressions-properties` endpoint, the UI is now able to pass a list of expressions to the server, which will return the unique properties present in that list.

This endpoint is used during the action saving process to build the input schema of the action based on the list of expressions contained in the mappings. Using this method allows removing all the logic from the application that calculates the action's input schema from the mappings and replaces it with a single HTTP call to the server.