admin: fix mapping rendering issue in `Action`
Currently, in the `Action` component, an exception occurs when you click
on the "Switch to transformation function" button, after mapping at
least one identifier and defining at least one mapping.

The exception is raised by React itself precisely when it tries to
remove the mapping list from the DOM to show in its place the
transformation's code editor.

The cause seems to be related to the fact that React tries to remove the
list using the `removeChild()` method on the parent node of the list,
but since the list has been automatically enclosed in a fragment, there
is no real node on which to call the `removeChild()` method.

This commit solves the problem by enclosing the mapping list within a
`div` element to give React a real parent node.