admin: check identifiers types
With this commit, an additional check is introduced on identifiers,
which now must comply with a specific type (for information on the
supported types see
https://github.com/open2b/chichi/wiki/Identity-Resolution).

Now, the property lists shown below the identifier inputs contain only
properties that have a type supported as an identifier. If the user
manually writes the name of a non-compliant property, the error message
"Type {type} is not supported as identifier" is displayed below the
input; trying to save, the operation is not completed and an additional
error is shown via a toast notification.

Given the increasing amount of logic related to identifiers, an
`identifiers` module has been added in `lib` in which all the validation
and transformation logic for identifiers has been inserted.

Furthermore, the logic that deals with adding, removing, moving, and
updating identifiers has been abstracted in the custom hook
`useIdentifiersMapping` to simplify the `IdentifiersMapping` component.