admin: fix error when trying to login by pressing enter
Currently, if you fill in the inputs in the login form and press the
"Enter" key on the keyboard to submit the form, before removing focus
from the inputs, a Javascript error occurs.

The problem arises because the values sent to the server for the login
are updated in the client state only when the user removes focus from
the inputs. Hence, if a user enters the email, fills in the password,
and presses "Enter" without defocusing, the server receives an empty
password resulting in an error.

This commit addresses the issue by changing the DOM event that triggers
the values updates in the state, switching from "SlChange" to "SlInput".
This ensures that the values are updated each time the user types.