admin: fix error on login/logout after toast notification activation
Currently, if you log in or out after the toast notification has been
activated, an application rendering error occurs.

The issue stems from the toast notification being unmounted and then
remounted in the DOM. Since it doesn't have a defined parent element
(the parent is a React Fragment), JavaScript throws an exception during
the "parentNode.insertBefore" operation.

This commit addresses the problem by wrapping the toast notification in
a `div`.