admin: use `createBrowserRouter` to define the routing
In the new versions of React Router, the `createBrowserRouter` function
has been introduced, which allows the application routing to be defined
using a standard Javascript object rather than using JSX.

This function also enables the incorporation of additional information
into the various routes through the use of the `handle` field, with this
information then available for use within the components.

Recognizing that the use of `handle` enables improvements to the
application, this commit transitions the definition of routing from
components to an object and utilizes `createBrowserRouter`.