admin: fix error when opening events actions with `null` event type
Currently, the UI expects all action types targeting `Events` to also
have an event type. However, some action types that operate generically
across multiple types of events are returned by the server with the
event type field set to `null`.

This difference in type interpretation is evidenced by the fact that
when editing an action targeting `Events`, to retrieve the corresponding
action type, the UI relies precisely on the value present in the
`EventType` field.

As a result, if the action has an event type equal to `null`, the UI
selects as the corresponding action type the first action type that has
`null` as the value in the `EventType` field. This sometimes leads to
selecting the wrong action type and thus causes errors in the rendering
of the action route.

This commit solves this issue by introducing an additional check to
ensure that the correct action type is always selected.