admin: remove transformation's properties
This commit eliminates the need within the UI to manually define the input and
output properties of an action's transformation.

Consequently, the transformation editor is displayed without the two side
columns containing the selected properties. The two dialogue screens that used
to show the list of selectable properties have been removed.

The transformation schemas are now set upon action save. They comprise a list
including the names of all the top-level properties contained in the action
schemas. These were as they existed when the user first accessed the creation
screen.