apis/mappings: ensure properties returned by Python are not identifiers
This commit changes the method `Mapping.Apply` to ensure that the Python
transformation function does not return any property value that has
already been mapped by the mappings (i.e., it is an identifier for the
action).

Note that this check can only be performed at runtime because at compile
time we no longer know which property will be returned by Python.