apis: don't require the mapping of the "id" property anymore
Currently, for file and database import actions, it is mandatory to map
the "id" property, but then such property is deleted by the server (it
is no longer used); so, this commit removes the needing of mapping the
"id" property for files and databases.

See the issue #221 for more details.

Closes #221.