apis: make state-changing methods return after the state is changed
Before this commit, an API method could change the state in the database
but would return before the state change was fully propagated. This
commit ensures that the state is propagated to the current node before
returning.

However, it does not guarantee that the change is propagated to the
other nodes before returning.

For #193