admin: improve `Action` navigation
Currently, the action editing and creation screen is a subcomponent of
the actions list. The screen itself does not correspond to any URL and
if the browser is refreshed during viewing, it will return to the
actions list view.

With this commit, a set of URLs is defined that will allow the user to
navigate directly to the action screen, without going through the
actions list. The URLs in question are:
- `/admin/connections/:id/actions/edit/:action`
- `/admin/connections/:id/actions/add/event/:eventType`
- `/admin/connections/:id/actions/add/event`
- `/admin/connections/:id/actions/add/:actionTarget`

Based on the URL, the UI will be able to understand the type of screen
to display, which will be respectively:
- editing of an existing action with id `:action`
- creation of an action based on the action type that targets `Events`
  and has `:eventType` as the event type
- creation of an action based on the action type that targets `Events`
  and has `null` as the event type
- creation of an action based on the action type that targets
  `:actionTarget`, which can be `users` or `groups`

Despite the action screen now being located at a different URL than the
actions list, the opening and closing animations have been maintained,
as well as the animation that occurs upon creating a new action.

Moreover, prior to this commit, the UI would manipulate the browser's
history API to ensure that if the browser's "back" button was clicked
during the display of the action screen, the user was redirected to the
actions list, even though this was not the previous page but only the
parent component.

This functionality has now been removed because, with the actions list
and the action screen being two completely separate routes, the browser
itself manages the navigation as the user expects. Consequently,
complexity has been removed from the UI, and a yet-to-be-investigated
bug has been resolved, which always redirected the user back to the
actions list when clicking on the browser's "back" button.