apis,apis/state: fix state load and listeners
This commit fixes a bug, introduced in commit 99e976a, where the
listeners on the state may not receive the first events received after
loading the state. This happens because, mistakenly, the listeners are
added after loading the state instead of before, as it should be.

To fix this, this commit separates the creation of the state, using the
new `state.New` function, from the loading of the state.