apis: reload only the schemas for users and groups
Currently, the `(*Workspace).ReloadSchema` method reloads the schema for
each table. However, only the schemas for users and groups are used.
This change ensures that only these schemas are reloaded.

Additionally, it should be noted that the comment:

      // The schema of the "events" table is hardcoded in the file
      // "apis/events/schema.go".

was incorrect, as the "apis/events/schema.go" file contains the schema
of the enriched event, not the data warehouse table's schema.