admin: reset `newActionID` to prevent unnecessary animation replay
Currently, after creating a new action, the corresponding row in the
action list is displayed with an entrance animation.

However, if another action is removed after the animation ends, the same
animation is played again.

This behavior is due to `newActionID`, the `ref` used to track the ID of
the potentially new action created, never being reset. Therefore, the
new repaints of the list caused by the removal of a row still consider
the ID to display the animation.

This commit resolves the issue by resetting `newActionID` before
removing a row to avoid triggering a new repaint.