admin: center the toast notification based on the URL
Currently, the admin app layout includes a sidebar for the main
navigation.

The toast notifications displayed after the UI events are centered
considering this sidebar to maintain a consistent aesthetic: they aren't
centered on the entire viewport, but only on the central content,
excluding the sidebar.

This approach, however, causes an unpleasant misalignment in those
routes that don't have the sidebar (like the login route or the action
creation/editing route) as notifications aren't centered on the entire
viewport and appear shifted to the right.

Previously, a setter method was used to set the `isFullscreen` variable,
which was considered for aligning the notifications based on the page.
This approach introduced complex logic and potential race conditions.

Having defined a new route specifically for action creation/editing
(previously a subcomponent of the action list) every page now
corresponds to a unique URL.

As a result, the logic that determines if the current route has a
sidebar can now be extracted from the components and inferred directly
from the URL.

This commit implements this new logic and centers the toast
notifications based on the current URL.