trace-events-script: make `identify` call include previously set traits
Prior to this commit, the `identify` call would only send the traits
passed as arguments. With this change, traits passed to the `identify`
call are now stored in the local storage, and upon subsequent `identify`
calls, these stored traits are retained and merged with the new traits
provided in the call. As a result, all traits are sent along with the
event.

Additionally, setting a trait to `undefined` will remove that trait from
the stored data.

Furthermore, this update introduces the `chichianalytics.user().traits`
method, to retrieve the stored traits and set the traits as a whole.