admin: fix `undefined` calls in `Action`
Currently, when opening the action creation/editing screen, a call is
made to the endpoint `admin/api/connections/:id/action-schemas/:target`
to retrieve the schemas of the action type.

However, the action type target, which is necessary to retrieve schemas,
is determined based on the browser's URL after the first render of
`ActionWrapper`. For this reason, during the first render of `Action`,
the action type will not be available yet, causing a call to
`admin/api/connections/:id/action-schemas/undefined` which returns a 404
error that is logged in the console.

This commit introduces a loading state within the `Fullscreen`
component, which contains `Action`. With this additional state, the
screen will display a spinner and will not render until the action type
is determined and available.