admin: always clear actions fetch interval when  opening `Action`
Currently, in the actions list, HTTP requests are made at intervals to
keep actions updated and track execution statuses.

This interval is suspended when opening the action editing/creation
screen, and resumed when the same screen is closed.

However, if the action screen is opened directly via URL (i.e., without
passing through the actions list), the interval is not suspended, and
HTTP calls continue in the background. This is due to the fact that
`ActionWrapper` is a subroute of `ConnectionActions`.

With this commit, this issue is resolved by moving the logic specifying
whether the action screen is open directly into an effect in
`ActionWrapper`, rather than in the callbacks of `ConnectionActions`.