admin: add a global loading state
This commit introduces a loading state at the application level. A
full-screen spinner is now displayed while the essential data to run the
application are being loaded.

The subcomponent tree will only be rendered once this loading is
finished.

Thanks to this change, it is possible to remove some defensive
programming logic that had to check whether data existed before
proceeding with the loading of subcomponents.

Additionally, this change fix some bugs, like for instance various 404
errors caused by the usage of data which was partially retrieved or
still not retrieved at all.

Note that the spinner is displayed for a minimum of 300ms and only if
the fetch operations take more than 100ms. Otherwise, no spinner is
shown, to avoid content flashing and to improve usability.