admin: fix styling issue in `ConnectorKeyValue`
Commit #4052181 renamed `row` class in the `ConnectorKeyValue` component
to `keyValueRow`. This was done as `row` was too generic and could
conflict with other styles.

However, it introduced a style issue as the class name was only changed
in the header row and not in the editable rows.

This new commit fixes the issue by renaming the missing classes.