admin: fix path used for redirecting to create a new storage connection
When creating a new connection based on a `File` type connector, if no
`Storage` type connection exists to associate it with, a "Create one..."
button appears on the configuration screen. This button redirects to the
list of connectors to first create a `Storage` type connection.

However, the path used for the redirect is of the form
`/connectors?role=${connectionRole}`, and since the `redirect` function
takes the path and concatenates it with `/admin/`, the result is a
redirect to `/admin//connectors?role=${connectionRole}` (with two
slashes). This causes the "not found" page to be displayed.

This commit solves the problem by removing the slash from the path
passed to `redirect`.