admin: disable the `Remove` button in the last identifiers association
Currently, in the identifiers mapping, it is not possible to remove the
last remaining association as there must always be at least one
identifier. If you try to remove the last association, it is not
actually removed and instead the error `You must define at least one
mapping` is shown.

With this commit, this mechanism is strengthened by directly disabling
the `Remove` button, which is no longer clickable if it is related to
the last remaining association.