admin: use `.jsx` extension where needed
Currently, in the ESBuild configuration, the `jsx` loader is used
indiscriminately for all Javascript files, including those with the
standard `js` extension.

Indeed, all the files in the `admin` directory have the `.js` extension,
even when they contain JSX.

This commit renames all files containing JSX, changing their extension
from `.js` to `.jsx`, and removes the option in the ESBuild
configuration that uses the `jsx` loader on all files.

The purpose of this change is to explicitly indicate which files contain
JSX and which do not, in addition to avoiding the unnecessary extra work
done by the `jsx` loader in the case of files that do not contain any
JSX.