admin: modify the ESBuild `Build` configuration
This commit modifies the configuration used to build the admin
application through ESBuild. In particular:

- React's development mode has been activated. This allows for
  additional errors, warnings and informations to be logged in the
  browser's console.

- The bundle code is no longer minified. This allows the bundle to be
  viewed with the original identifiers and formatting. In addition, the
  component hierarchy displayed in the `components` tab of React
  Developer Tools now shows the original names of the components, state
  variables and props, and not the minified names.

- Sourcemaps have been activated. Now, when an error occurs in
  Javascript, the browser requests a sourcemap from the server, with
  which is able to display the exact original file (thus also
  maintaining the type notations in the case of Typescript) and the
  exact line where the error occurred.

Note that to add the sourcemaps, a modification was made to the
`serveWithESBuild` function which will now also have to serve client
requests for files that end with the `.map` extension.

These build options will need to be reviewed before going into
production and for this purpose, TODOs have been inserted on the lines
of the options in question.