admin: improve error message in `serveWithESBuild`
Currently, within `serveWithESBuild`, if a request with an unsupported
`Content-Type` arrives from the client, an `Internal Server Error` is
returned and the error is logged.

Actually, since it's the client making a request for something the
server does not support, this error should be a `Bad Request`.

This commit therefore modifies the management of this error by returning
a `Bad Request` to the client and avoiding logging the error.