admin: remove the usage of the `CHICHI_DEBUG_UI` environment variable
Currently, the server checks the environment variable `CHICHI_DEBUG_UI`,
and if it is present and set to `true`, in case of a `Bad Request`
error, it will send an additional error message containing more
information to the client, which will be shown in a toast notification.

With this commit, it has been decided to no longer check the environment
variable and to always send the error message.

Therefore, this commit establishes the criterion that the information of
any `Bad Request` errors should always be shown to users, as these can
be helpful during support for debugging purposes. The commit also
highlights the importance of checking that there is no information in
the `Bad Request` error messages that should be hidden from the user, as
from now on, these would be shown both during development and in
production.

This commit also removes further references to the `CHICHI_DEBUG_UI`
environment variable in the tests, in the README, and in the UI.