all: add modules vendoring to every module in the repository
After a discussion with @gazerro, we decided to introduce the module
vendoring in Chichi to prevent future problems (as, for example, third
party modules that may no longer be maintained and disappear).

So, this commits adds the module vendoring to the main module and to
every submodule in the repository, and also update the script
`commit/commit.go` to perform such vendoring.