admin: implement a user-friendly error screen
This commit introduces the `RootError` component, which is passed to
`react-router` to display an error screen that is more aligned with the
design of the application.

This screen is shown every time an unhandled exception is raised within
Javascript. The additional informations, mapped to the source code via
sourcemap, are automatically logged in the console.