all: remove vendor from every module of the repository
The commit 03288cb04cf15c72e8d8a741f0541ad6948c927a added the vendor to
every module of the repository. Unfortunately, this lead to an error
with the editor JetBrains Goland, which seems to be unable to detect
dependencies outside the current module when there is a "vendor"
directory in it, making impossible to develop the connectors (as they
have a dependency with the main module "chichi").

So, this commit removes the vendor from the repository.