admin: fix `null` reference error on `toastRef` during initial render
Currently, if a logged-in user manually inputs the `/admin/` URL into
the browser, the application encounters an error.

When the user navigates to this URL while logged in, instead of being
shown the login screen again, they are redirected to the `/connections/`
URL. There, they are presented with a map of installed connections.

The `redirect` method, which takes care of this operation, hides any
opened toast notification by utilizing `toastRef.current.hide()`.
However, during the application's first rendering, `toastRef` will not
yet have been assigned and its value, accessible through
`toastRef.current`, will be `null`. Thus, invoking the `hide( )` method
on a `null` value leads to the error.

This commit fixes the error by implementing a check on the value of
`toastRef.current`. It avoids closing the toast notification if it has
not yet been initialized.