admin: add `SettingsForm` to unify connection and connector settings
With this commit, the `SettingsForm` component is implemented, which
aims to avoid code repetition for the connection and connector settings
forms.

In particular, `SettingsForm` defines the `SettingsContext`, the markup
and the styles used by the settings.

As a side effect, this change fixes some stylistic inconsistencies that
had gradually formed between the connection settings page and the
connector settings page.