admin: fix replication of first identifier association in `Action`
Currently, if a new action is created immediately after creating another
one, the first pair of identifier inputs will already be filled with the
values used in the previously created action.

This behavior also occurs when actions are created in sequence but on
different connections (if both have identifiers) and typically leads to
errors during saving, because the values used in one action may not
exist in the schema of the action type of the next action.

When a new action is created, a default action is automatically
generated for the component to display the screen (this happens in the
`computeDefaultAction()` function).

In this generated action, if the action type supports identifiers, the
`Identifiers` field of the action is also pre-filled with a default
value contained in the `DEFAULT_IDENTIFIERS_MAPPING` constant.

The bug in question arises because the assignment of the constant does
not occur with a deep copy: instead, the `Identifiers` field will
receive a reference to `DEFAULT_IDENTIFIERS_MAPPING` and therefore when
the user modifies the identifiers of the action, they will also end up
modifying the underlying constant.

Subsequently, when creating a new action, when generating the default
action, `DEFAULT_IDENTIFIERS_MAPPING` will be used again, but this time,
instead of the default value, the value of the constant will be the one
modified by the previous use.

This commit solves the problem by ensuring to pass a deep copy of the
constant instead of the reference, and therefore avoiding changes to the
underlying value.