admin: change `Remove` button disable logic in identifiers lists
Currently in the identifiers lists, the "Remove" button is disabled on
the first pair of inputs.

In order to improve the usability of the section, with this commit, the
"Remove" button is disabled on the first pair of inputs only if it is
the only pair in the list and only if both inputs are empty.

If there is only one pair of inputs, but if at least one of the two
inputs is filled, the "Remove" button will still be enabled. Clicking it
will automatically clear both inputs.