admin: update actions mapping after identifier removal
Currently, when an identifier is removed from the identifiers list in
the action editing screen, the related destination property is returned
to the mapping list, and the property that was mapped to the identifier
is also returned in the mapping as its input property.

If the identifier removal occurs after changing the property that was
mapped to it, the identifier will be returned to the mapping list with
the old mapped property, i.e., the one that was present when the action
was saved and not the one present after the change.

Given that this behavior could be confusing for the UI user, this commit
ensures that when an identifier is removed, the corresponding mapping is
also removed. Therefore, once the identifier is returned to the mapping
list as a destination property, no input property will be mapped to it.