apis/warehouses: remove the `Warehouse.Query` method
In the `Warehouse` interface, there are methods that accept arguments
depending on the type of data warehouse. These methods should be
replaced with ones that have arguments independent of the specific data
warehouse type, such as the `Select` method.

This commit, therefore, begins by removing the `Query` method, which is
currently not in use.