all: make sure each workspace has its own Redis database
Currently, there is a single Redis database for all the workspaces. This
change makes sure that each workspace has its own database.

It removes the Redis settings from the configuration file and adds
two API methods to connect and disconnect a Redis database:

   * POST /api/workspace/connect-redis
   * POST /api/workspace/disconnect-redis

Additionally, two command are added to the CLI:

   * ./chichi-cli connect-redis redis.json
   * ./chichi-cli disconnect-redis