admin: add placeholders to action mapping
A property contained in the output schema of an action type can have a
"placeholder" value, which can be used as the default value of the
associated input property in the action's mapping.

This commit introduces exactly this additional functionality by using
placeholders when calculating initial mappings, upon opening the action
creation screen, or when switching from the transformation function to
the mappings.