admin: restore scroll on grids
Scrolling in the grids had been disabled via CSS in previous commits.
This led to an issue with the user grid, which is designed to be
'embedded' within the page with a limited height, and also can contain a
significant list of columns. This typically caused an overflow on most
viewports.

This commit restore the scroll feature to address the issue.