admin: fix `Schedule` dropdown opening after action creation
Currently, if you create a new action and try to click the `Schedule`
button before reloading the page, the dropdown list with the time
periods available for the scheduler will not be displayed.

The issue arises from a conflict between Shoelace and the entrance
animation of the newly created action.

This commit solves the problem by slightly modifying the animation to
accommodate the positioning algorithm used internally by Shoelace.