admin: fix logo mismatches in the connections map
Currently, when a new connection is created, it is displayed with an
entrance animation in the connections map. However, an issue arises with
the logo of the newly created connection: it does not reflect the logo
actually associated with its connector, but instead, it shows the logo
that was previously rendered for the connection in the same position on
the map.

This behavior stems from the fact that the image in question is inserted
into the DOM only after the first rendering of the container component
using a `useEffect` hook with an empty dependency array. Thus,
subsequent renderings will not cause the modification of the logo, which
will remain the one previously set.

This commit resolves the issue by forcing the reloading of the
connection logo with every rendering of the container component.