apis: fix update of user index on Redis
Currently, when updating a user, only the properties of that specific
user are being updated on Redis. As a result, any other properties that
might have been imported previously or from another action are being
unintentionally deleted from the index.

This commit fixes this bad behavior, integrating the properties which
are currently on Redis with the incoming properties.

Also, it improves some documentation and code formatting.